fix: enhance SQL query type handling in upsert_current_rate
- Updated the upsert_current_rate SQL query to explicitly cast buy_price and sell_price to DECIMAL for improved data integrity and consistency during database operations.

+++ app/core/database_optimized.py
@@ -69,7 +69,7 @@ PREPARED_QUERIES = {
     "upsert_current_rate": """
         INSERT INTO current_rates (exchange_code, currency_pair, buy_price, sell_price, avg_price,
                                  variation_24h, volume_24h, source, last_update, market_status)
-        VALUES ($1, $2, $3, $4, ($3 + $4) / 2, $5, $6, $7, NOW(), 'active')
+        VALUES ($1, $2, $3::DECIMAL, $4::DECIMAL, ($3::DECIMAL + $4::DECIMAL) / 2, $5::DECIMAL, $6::DECIMAL, $7, NOW(), 'active')
         ON CONFLICT (exchange_code, currency_pair) 
         DO UPDATE SET 
             buy_price = EXCLUDED.buy_price,
